Split up CommunityVideos
