+++ app/javascript/components/track/approaches-elements/community-videos/index.ts
@@ -1,0 +1,2 @@
+export { CommunityVideo, CommunityVideosFooter } from './CommunityVideo'
+export { CommunityVideos } from './CommunityVideos'
