+++ app/javascript/components/track/approaches-elements/index.ts
@@ -1,6 +1,6 @@
 export { DiggingDeeper, ApproachIntroduction } from './DiggingDeeper'
 export { ApproachExamples } from './ApproachExamples'
 export { SectionHeader } from './SectionHeader'
-export { CommunityVideos, CommunityVideosProps } from './CommunityVideos'
+export { CommunityVideos } from './community-videos'
 export { NoContentYet, NoIntroductionYet } from './no-content-yet'
 export { ApproachSnippet } from './ApproachSnippet'
